backend/requirements.txt
asgiref==3.11.0
Django==4.2.26
django-cors-headers==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
python-dotenv==1.2.1
sqlparse==0.5.4
typing_extensions==4.15.0
